// src/app/views/canopy-process/powder-coating-maker/Model/powdercoatingprcPartDts.ts
export interface IpowdercoatingPrcPartDts {
  SelectPC: boolean;
  KVA: string;
  ModelCPType: string;
  KitDesc: string;
  Sqft: number;
  BatchQty: number;
  BatchBalQty: number;
  TotSqft: number;
  PrcQty: number;
  PrcSqft: number;
  PlanCode: string;
  PlanDt: string;
  KitCode: string;
  PfbCode: string;
  EDt: string;
  BOMCode: string;
  ProductCode: string;
  Rate: number;
  GroupPfbCode: string;
  CatID: string;
}